Dry up logging tests
