+++ keystone_api/tests/logging/test_requests.py
@@ -1,76 +1,12 @@
 """Function tests for the `/logs/requests/` endpoint."""
 
-from rest_framework import status
 from rest_framework.test import APITestCase
 
-from apps.users.models import User
-from tests.utils import CustomAsserts
+from .common import LoggingPermissionTests
 
 
-class EndpointPermissions(APITestCase, CustomAsserts):
-    """Test endpoint user permissions.
-
-    Endpoint permissions are tested against the following matrix of HTTP responses.
-
-    | User Status                | GET | HEAD | OPTIONS | POST | PUT | PATCH | DELETE | TRACE |
-    |----------------------------|-----|------|---------|------|-----|-------|--------|-------|
-    | Unauthenticated User       | 403 | 403  | 403     | 403  | 403 | 403   | 403    | 403   |
-    | Authenticated User         | 403 | 403  | 403     | 403  | 403 | 403   | 403    | 403   |
-    | Staff User                 | 200 | 200  | 200     | 405  | 405 | 405   | 405    | 405   |
-    """
+class EndpointPermissions(LoggingPermissionTests, APITestCase):
+    """Test endpoint user permissions."""
 
     endpoint = '/logs/requests/'
     fixtures = ['testing_common.yaml']
-
-    def setUp(self) -> None:
-        """Load user accounts from testing fixtures."""
-
-        self.staff_user = User.objects.get(username='staff_user')
-        self.generic_user = User.objects.get(username='generic_user')
-
-    def test_unauthenticated_user_permissions(self) -> None:
-        """Verify unauthenticated users cannot access resources."""
-
-        self.assert_http_responses(
-            self.endpoint,
-            get=status.HTTP_403_FORBIDDEN,
-            head=status.HTTP_403_FORBIDDEN,
-            options=status.HTTP_403_FORBIDDEN,
-            post=status.HTTP_403_FORBIDDEN,
-            put=status.HTTP_403_FORBIDDEN,
-            patch=status.HTTP_403_FORBIDDEN,
-            delete=status.HTTP_403_FORBIDDEN,
-            trace=status.HTTP_403_FORBIDDEN
-        )
-
-    def test_authenticated_user_permissions(self) -> None:
-        """Verify authenticated users are returned a 403 status code for all request types."""
-
-        self.client.force_authenticate(user=self.generic_user)
-        self.assert_http_responses(
-            self.endpoint,
-            get=status.HTTP_403_FORBIDDEN,
-            head=status.HTTP_403_FORBIDDEN,
-            options=status.HTTP_403_FORBIDDEN,
-            post=status.HTTP_403_FORBIDDEN,
-            put=status.HTTP_403_FORBIDDEN,
-            patch=status.HTTP_403_FORBIDDEN,
-            delete=status.HTTP_403_FORBIDDEN,
-            trace=status.HTTP_403_FORBIDDEN
-        )
-
-    def test_staff_user_permissions(self) -> None:
-        """Verify staff users have read-only permissions."""
-
-        self.client.force_authenticate(user=self.staff_user)
-        self.assert_http_responses(
-            self.endpoint,
-            get=status.HTTP_200_OK,
-            head=status.HTTP_200_OK,
-            options=status.HTTP_200_OK,
-            post=status.HTTP_405_METHOD_NOT_ALLOWED,
-            put=status.HTTP_405_METHOD_NOT_ALLOWED,
-            patch=status.HTTP_405_METHOD_NOT_ALLOWED,
-            delete=status.HTTP_405_METHOD_NOT_ALLOWED,
-            trace=status.HTTP_405_METHOD_NOT_ALLOWED
-        )
